Move compaction inside libocto
